+++ tools/spec-gen-sdk/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release
 
+## 2025-03-14 - 0.3.0
+
+- Removed breaking change label artifact related code
+- Refactored loggings for pipeline scenarios
+
 ## 2025-03-11 - 0.2.2
 
 - Saved label artifacts in separate folders